feat:Integrando Api de terceiro para busca de exercicios e seus dados

+++ src/main/java/com/Gymlog/Repository/WorkoutSetsRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface WorkoutSetsRepository extends JpaRepository<WorkoutSetsEntity, Integer> {
+public interface WorkoutSetsRepository extends JpaRepository<WorkoutSetsEntity, Long> {
 }
